HTML how to format blog post
<!DOCTYPE html> (this type of this document is HTML) 
<html lang="en"> (use html elements to define the structure of this web page)
(So then, down here I can start creating what I called container tags. )
<head>
<title>Implementation of HTML elements</title>
<meta charset="UTF-8">
</head>
<body>
<h1> Implementation of HTML elements</h1>

(In our implementation of the code, paragraphs in HTML are set using a <p> tag </p>. The text contents for the paragraph are placed between the tags. Paragraphs elements are used to format many of the texts that are longer than one sentence.)
Example of paragraph implementation is:
<p>this is a paragraph</p>
(down here I can start typing what I want in there) 


<p> This is a paragraph implementation to show how 
	a text behaves in an html web pages.
	It is placed between the p tags and can be formatted the way we want(Jackson Nguyen)</p>

This is implemented as an unordered list in the HTML web pages. The list of items in the element is bulleted. The primary tag for the bulleted list is <ol>, and the individual items are implemented using the <li> tag. 
This has been coded as shown below:

<ul>
	<li> Bullet List Item 1</li>
	<li> Bullet List Item 2</li>
	<li> Bullet List Item 3</li>
	<li> Bullet List Item 4</li>
</ul>
<ol>
	<li> Bullet List Item 1</li>
	<li> Bullet List Item 2</li>
	<li> Bullet List Item 3</li>
	<li> Bullet List Item 4</li>
</ol>
<table>  
	<tr><th>SN</th><th>Name</th><th>School</th></tr>  
	<tr><td>1</td><td>James A </td><td>633, ATL</td></tr>  
	<tr><td>2</td><td>N.Smith </td><td>300, MN</td></tr>  
	<tr><td>3</td><td>Adam Smith </td><td>40, TR</td></tr>  
	<tr><td>4</td><td>Reece Adams </td><td>520, ATL</td></tr>  
</table>
	<ul> 
	  <li><a href="#paragraph-id">1 Pointer Header For Paragraph</a> 
	  <ul> 
		<li><a href="#unordered-list">1.1 Unoredered List</a></li> 
		<li><a href="#ordered-list">1.2 Ordered List</a></li> 
	  </ul> 
	</li> 
	<li><a href="#table-data">2 Data Table</a></li> 
	</ul> 
</body>
</html>